requirements.txt
# Nexus - Personal AI Agent

# Core
pyyaml>=6.0

# Model clients
anthropic>=0.25.0
openai>=1.0.0
httpx>=0.27.0

# MCP support (optional - install MCP servers separately)
# mcp>=1.0.0

# Office tools (optional)
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
pypdf>=4.0.0

# Email (optional)
# imap-tools>=1.0.0

# Browser automation (optional)
selenium>=4.15.0
playwright>=1.40.0

# CLI enhancements
prompt_toolkit>=3.0.0
rich>=13.0.0
typer>=0.9.0